+++ src/main/resources/templates/fragments/header.html
@@ -22,10 +22,28 @@
             <th:block th:if="${user == null}">
                 <a href="/signin" class="login-btn">로그인</a>
             </th:block>
+            <!--
             <th:block th:if="${user != null}">
                 <div class="username" id="username" onclick="toggleDropdown()" th:text="${user.usernick}">사용자이름</div>
                 <div class="dropdown" id="dropdown-menu">
                     <a href="/my">내 정보</a>
+                    <a href="/api/users/logout">로그아웃</a>
+                    <a href="/faq">고객센터</a>
+                </div>
+            </th:block>-->
+            <th:block th:if="${user != null}">
+                <!-- 일반 사용자 또는 관리자 구분 -->
+                <div class="username" id="username" onclick="toggleDropdown()" th:text="${user.usernick}">사용자이름</div>
+                <div class="dropdown" id="dropdown-menu">
+                    <th:block th:if="${user.usernick == 'OMGAdmin'}">
+                        <a href="/admin/userboard">사용자 관리</a>
+                        <a href="/admin/reviewboard">게시글 관리</a>
+                        <a href="#">공지사항 관리</a>
+                    </th:block>
+                    <th:block th:if="${user.usernick != 'OMGAdmin'}">
+                        <a href="/my">내 정보</a>
+                        <a href="/faq">고객센터</a>
+                    </th:block>
                     <a href="/api/users/logout">로그아웃</a>
                 </div>
             </th:block>
