+++ src/main/resources/templates/main/faq.html
@@ -100,6 +100,34 @@
                 <div class="answer">
                     <p>등록한 리뷰는 웹 상단 내 정보 > 내 여행 후기에서 직접 수정/삭제가 가능합니다.</p>
                 </div>
+                <h3>[회원정보] 회원탈퇴는 어떻게 하나요?</h3>
+                <div class="answer">
+                    <p>회원탈퇴는 아래 방법을 통해 가능하며, 반드시 주의사항을 확인 후 진행해 주시기 바랍니다. <br/>
+                        [일반회원] <br/>
+                        - OMG 웹 > 내 정보 > 내정보 관리 > 회원탈퇴 <br/>
+                        *주의사항 <br/>
+                        · 회원탈퇴 시 계정 정보는 복구가 불가합니다. <br/>
+                        · 등록된 리뷰는 자동으로 삭제되지 않습니다. 삭제를 원하실 경우 탈퇴 전 개별적으로 삭제해 주시기 바랍니다.</p>
+                </div>
+                <h3>[회원정보] 비밀번호를 잊었어요.</h3>
+                <div class="answer">
+                    <p>[비밀번호 찾기]
+                        ※ 이메일주소 가입만 해당 <br/>
+                        1. 로그인 > 비밀번호 찾기 <br/>
+                        2. 가입 시 등록한 이메일 주소 입력 <br/>
+                        3. 이메일 주소로 발급된 새로운 비밀번호로 로그인 진행 <br/>
+                        4. 로그인 후 새로운 비밀번호 재설정 필수 <br/>
+                        <br/>
+                        위 방법으로도 이용에 불편을 겪으실 경우 OMG 고객센터로 문의해주시길 바랍니다.</p>
+                </div>
+                <h3>[채팅] 채팅을 어떻게 시작하나요?</h3>
+                <div class="answer">
+                    <p>채팅은 일정에 참여를 하게 되면 시작할 수 있습니다.</p>
+                </div>
+                <h3>[채팅] 채팅이 제대로 작동하지 않아요.</h3>
+                <div class="answer">
+                    <p>채팅이 작동하지 않을 경우, 브라우저의 캐시를 지우고 페이지를 새로고침 해보세요. 문제가 계속되면 OMG 고객센터로 문의해주시기 바랍니다.</p>
+                </div>
             </div>
 
             <!-- 이용문의 -->
